remove u32 error

+++ protocol/src/commands/deliver.rs
@@ -94,7 +94,7 @@ impl Encoder for DeliverCommand {
             .iter()
             .fold(0, |acc, message| acc + 1 + message.encoded_size());
 
-        writer.write_u32::<BigEndian>(size as u32)?;
+        writer.write_u32::<BigEndian>(size)?;
         self.trailer_length.encode(writer)?;
         self.reserved.encode(writer)?;
